Refactor game manager classes into separate files
