+++ server/socket/gameManager/Game.js
@@ -1,3 +1,5 @@
+const Player = require('./Player');
+
 class Game {
   constructor(roomname, username, deck) {
     this.name = roomname;
@@ -16,13 +18,8 @@ class Game {
 
   addPlayer(username) {
     this.playerCount += 1;
-    this.players.push({
-      username: username,
-      isCzar: false,
-      points: 0,
-      cards: [],
-      playerNum: this.playerCount
-    });
+    let player = new Player(username, this.playerCount);
+    this.players.push(player);
   }
 
   shuffle() {
@@ -101,43 +98,4 @@ class Game {
 
 }
 
-class GameManager {
-  constructor() {
-    this.games = {};
-    this.rooms = [];
-  }
-
-  createGame(roomname, username, deck) {
-    const newGame = new Game(roomname, username, deck);
-    this.games[roomname] = newGame;
-    this.rooms.push({
-      name: roomname,
-      createdBy: username
-    });
-  }
-
-  addPlayer(roomname, username) {
-    if (this.games[roomname]) {
-      this.games[roomname].addPlayer(username);
-    } else {
-      console.log('game room not found, did not add player');
-    }
-  }
-
-  getRoom(roomname) {
-    if (this.games[roomname]) {
-      return this.games[roomname];
-    } else {
-      console.log('game room not found, did not add player');
-    }
-  }
-
-  closeRoom(roomname) {
-    // remove room from list of rooms
-    // when a room creator closes a room to start a game
-  }
-}
-
-
-
-module.exports = new GameManager();
+module.exports = Game;
